Solve the board with a depth first search
The code is pretty horrible and is need of refactoring.

It relies on returning over 162 bytes of Board several times in the
depth_first_search() function. This probably doesn't happen too often
but still feels horribly wrong.

EDIT: the profiler confirms that this does not happen often enough to burn
a significant amount of time. In fact, most of the time is spent in
solve_one_iteration() and a little under half of that time is spent calling
std::vector::push_back() :-/